% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundaries_dpt18}
\alias{boundaries_dpt18}
\title{Departments (2018)}
\format{
A data frame of class "sf" with 102 rows and 4 variables
\describe{
   \item{dpt18_name}{Department name}
   \item{dpt18_nuts3}{Department NUTS3 level code}
   \item{dpt18_code}{Department INSEE code}
   \item{geometry}{Multipolygon coordinates}
}
}
\source{
\url{https://www.data.gouv.fr/fr/datasets/contours-des-departements-francais-issus-d-openstreetmap/}
}
\usage{
boundaries_dpt18
}
\description{
A dataset containing 2018 department geometries and their associated names and codes.
}
\keyword{datasets}
