% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_delivered.R
\name{filter_delivered}
\alias{filter_delivered}
\title{Filter delivered kits}
\usage{
filter_delivered(df)
}
\arguments{
\item{df}{A dataframe of orders with a "tracking_status" variable}
}
\value{
A dataframe
}
\description{
Filter delivered kits
}
