% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_delivered.R
\name{filter_delivered}
\alias{filter_delivered}
\title{Filter delivered kits}
\usage{
filter_delivered(df, col_name = tracking_status)
}
\arguments{
\item{df}{A dataframe of Delivraide orders}

\item{col_name}{The column to use for filtering (default: tracking_status). Note: do not add quotes around the column name.}
}
\value{
A dataframe
}
\description{
This function filters from the dataset the orders that have been delivered, and removes those that have been cancelled.
  It does so by using the specified tracking variable, defaulting to "tracking_status", that is present in Delivraide data.
}
\examples{
# Load the sample orders data
data(sample_orders_data)

# Filter the delivered orders
delivered_orders <- filter_delivered(sample_orders_data)

# Print the filtered dataframe
print(delivered_orders)

}
