% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_delivered.R
\name{filter_delivered}
\alias{filter_delivered}
\title{Filter delivered kits}
\usage{
filter_delivered(df)
}
\arguments{
\item{df}{A dataframe of Delivraide orders}
}
\value{
A dataframe
}
\description{
This function filters from the dataset the orders that have been delivered, and removes those that have been cancelled.
 It does so by using the "tracking_status" variable that is present in Delivraide data.
}
